services/orchestrator/requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
PyYAML==6.0.2
httpx==0.27.2
sqlite-utils==3.37
websocket-client
requests
fastapi>=0.115
uvicorn[standard]>=0.30
websocket-client>=1.6
aiohttp>=3.9
PyYAML>=6.0
openai>=1.55.0
redis>=4.6.0
prometheus-client
